organize tokenizer & start expander
